+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added:
-- Weekly aggregation in story time charts
+- Count duration for entries spanning multiple days for each individual day
 
 ## [0.8.112] - 2022-07-19
 ### Changed: